Moved DwellingInstanceConstructor to a new file
